a) print an error if sendbug failed. b  add debug support to win32 mail interface:     "BK_MAIL_DEBUG=1 bk sendbug" will show you the smtp     message exchange.
bk: 3ab01624MYXudCcDVsKoEWsttapIQQ

+++ src/sendbug.c
@@ -36,8 +36,14 @@ sendbug_main(int ac,  char **av)
 				exit(1);
 			}
 			waitpid(pid, &status, 0);
-			unlink(bug);
-			printf("Your bug has been sent, thank you.\n");
+			if (status == 0) {
+				printf("Your bug has been sent, thank you.\n");
+				unlink(bug);
+			} else {
+				printf(
+				    "mailer failed: bug report saved in %s\n",
+				    bug);
+			}
 			exit(0);
 		    case 'e':
 			sprintf(buf, "%s %s", editor, bug);
